requirements.txt
# Core Dependencies
Flask==3.1.2
flask-cors==6.0.2

# Deep Learning & ML
torch==2.10.0
torchvision==0.25.0
open-clip-torch==3.2.0

# Vector Search
faiss-cpu==1.13.2

# Image Processing
pillow==12.1.0

# Numerical Computing
numpy==2.4.1
scipy==1.17.0
scikit-learn==1.8.0

# Utilities
tqdm==4.67.1
timm==1.0.24
huggingface-hub==1.3.4
safetensors==0.7.0

# Dependencies Management
setuptools==80.10.2
wheel==0.46.3
